+++ semana20/labook-projeto/template-labook/request.rest
@@ -1,10 +1,10 @@
-post http://localhost:3003/users/signup
+post http://localhost:3003/user/signup
 Content-Type: application/json
 
 {
-    "name": "Jamile",
-    "email": "jm@email.com",
-    "password": "123456"
+    "name": "kdfghj",
+    "email": "dgftyh@email.com",
+    "password": "12346"
 }
