+++ semana20/labook-projeto/template-labook/src/routes/UserRouter/userRouter.ts
@@ -1,0 +1,9 @@
+import express from "express";
+import { UserController } from "../../controller/UserController";
+
+export const userRouter = express.Router()
+
+const userController = new UserController();
+
+userRouter.post("/signup", userController.signup)
+
